--- apps/meteor/client/components/BurgerMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BurgerMenu';
