+++ apps/meteor/client/components/SidebarToggler/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SidebarToggler';
