+++ apps/meteor/client/views/room/Header/Omnichannel/VoipRoomHeader.tsx
@@ -5,7 +5,7 @@ import type { FC } from 'react';
 import React, { useCallback, useMemo } from 'react';
 import { useSyncExternalStore } from 'use-sync-external-store/shim';
 
-import BurgerMenu from '../../../../components/BurgerMenu';
+import SidebarToggler from '../../../../components/SidebarToggler';
 import { parseOutboundPhoneNumber } from '../../../../lib/voip/parseOutboundPhoneNumber';
 import type { RoomHeaderProps } from '../RoomHeader';
 import RoomHeader from '../RoomHeader';
@@ -30,7 +30,7 @@ const VoipRoomHeader: FC<VoipRoomHeaderProps> = ({ slots: parentSlot, room }) =>
 			...parentSlot,
 			start: (!!isMobile || currentRouteName === 'omnichannel-directory') && (
 				<HeaderToolbar>
-					{isMobile && <BurgerMenu />}
+					{isMobile && <SidebarToggler />}
 					{currentRouteName === 'omnichannel-directory' && <BackButton />}
 				</HeaderToolbar>
 			),
